+++ packages/cloud_firestore_odm/cloud_firestore_odm_generator/cloud_firestore_odm_generator_integration_test/lib/freezed.g.dart
@@ -8,6 +8,7 @@ part of 'freezed.dart';
 
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, require_trailing_commas, prefer_single_quotes, prefer_double_quotes, use_super_parameters, duplicate_ignore
+// ignore_for_file: type=lint
 
 class _Sentinel {
   const _Sentinel();
@@ -189,15 +190,15 @@ class _$PersonDocumentReference
     );
     final json = {
       if (firstName != _sentinel)
-        _$$_PersonFieldMap['firstName']!:
-            _$$_PersonPerFieldToJson.firstName(firstName as String),
+        _$$PersonImplFieldMap['firstName']!:
+            _$$PersonImplPerFieldToJson.firstName(firstName as String),
       if (firstNameFieldValue != null)
-        _$$_PersonFieldMap['firstName']!: firstNameFieldValue,
+        _$$PersonImplFieldMap['firstName']!: firstNameFieldValue,
       if (lastName != _sentinel)
-        _$$_PersonFieldMap['lastName']!:
-            _$$_PersonPerFieldToJson.lastName(lastName as String),
+        _$$PersonImplFieldMap['lastName']!:
+            _$$PersonImplPerFieldToJson.lastName(lastName as String),
       if (lastNameFieldValue != null)
-        _$$_PersonFieldMap['lastName']!: lastNameFieldValue,
+        _$$PersonImplFieldMap['lastName']!: lastNameFieldValue,
     };
 
     return reference.update(json);
@@ -220,15 +221,15 @@ class _$PersonDocumentReference
     );
     final json = {
       if (firstName != _sentinel)
-        _$$_PersonFieldMap['firstName']!:
-            _$$_PersonPerFieldToJson.firstName(firstName as String),
+        _$$PersonImplFieldMap['firstName']!:
+            _$$PersonImplPerFieldToJson.firstName(firstName as String),
       if (firstNameFieldValue != null)
-        _$$_PersonFieldMap['firstName']!: firstNameFieldValue,
+        _$$PersonImplFieldMap['firstName']!: firstNameFieldValue,
       if (lastName != _sentinel)
-        _$$_PersonFieldMap['lastName']!:
-            _$$_PersonPerFieldToJson.lastName(lastName as String),
+        _$$PersonImplFieldMap['lastName']!:
+            _$$PersonImplPerFieldToJson.lastName(lastName as String),
       if (lastNameFieldValue != null)
-        _$$_PersonFieldMap['lastName']!: lastNameFieldValue,
+        _$$PersonImplFieldMap['lastName']!: lastNameFieldValue,
     };
 
     transaction.update(reference, json);
@@ -580,29 +581,29 @@ class _$PersonQuery extends QueryReference<Person, PersonQuerySnapshot>
     return _$PersonQuery(
       _collection,
       $referenceWithoutCursor: $referenceWithoutCursor.where(
-        _$$_PersonFieldMap['firstName']!,
+        _$$PersonImplFieldMap['firstName']!,
         isEqualTo: isEqualTo != null
-            ? _$$_PersonPerFieldToJson.firstName(isEqualTo)
+            ? _$$PersonImplPerFieldToJson.firstName(isEqualTo)
             : null,
         isNotEqualTo: isNotEqualTo != null
-            ? _$$_PersonPerFieldToJson.firstName(isNotEqualTo)
+            ? _$$PersonImplPerFieldToJson.firstName(isNotEqualTo)
             : null,
         isLessThan: isLessThan != null
-            ? _$$_PersonPerFieldToJson.firstName(isLessThan)
+            ? _$$PersonImplPerFieldToJson.firstName(isLessThan)
             : null,
         isLessThanOrEqualTo: isLessThanOrEqualTo != null
-            ? _$$_PersonPerFieldToJson.firstName(isLessThanOrEqualTo)
+            ? _$$PersonImplPerFieldToJson.firstName(isLessThanOrEqualTo)
             : null,
         isGreaterThan: isGreaterThan != null
-            ? _$$_PersonPerFieldToJson.firstName(isGreaterThan)
+            ? _$$PersonImplPerFieldToJson.firstName(isGreaterThan)
             : null,
         isGreaterThanOrEqualTo: isGreaterThanOrEqualTo != null
-            ? _$$_PersonPerFieldToJson.firstName(isGreaterThanOrEqualTo)
+            ? _$$PersonImplPerFieldToJson.firstName(isGreaterThanOrEqualTo)
             : null,
         isNull: isNull,
-        whereIn: whereIn?.map((e) => _$$_PersonPerFieldToJson.firstName(e)),
+        whereIn: whereIn?.map((e) => _$$PersonImplPerFieldToJson.firstName(e)),
         whereNotIn:
-            whereNotIn?.map((e) => _$$_PersonPerFieldToJson.firstName(e)),
+            whereNotIn?.map((e) => _$$PersonImplPerFieldToJson.firstName(e)),
       ),
       $queryCursor: $queryCursor,
     );
@@ -622,29 +623,29 @@ class _$PersonQuery extends QueryReference<Person, PersonQuerySnapshot>
     return _$PersonQuery(
       _collection,
       $referenceWithoutCursor: $referenceWithoutCursor.where(
-        _$$_PersonFieldMap['lastName']!,
+        _$$PersonImplFieldMap['lastName']!,
         isEqualTo: isEqualTo != null
-            ? _$$_PersonPerFieldToJson.lastName(isEqualTo)
+            ? _$$PersonImplPerFieldToJson.lastName(isEqualTo)
             : null,
         isNotEqualTo: isNotEqualTo != null
-            ? _$$_PersonPerFieldToJson.lastName(isNotEqualTo)
+            ? _$$PersonImplPerFieldToJson.lastName(isNotEqualTo)
             : null,
         isLessThan: isLessThan != null
-            ? _$$_PersonPerFieldToJson.lastName(isLessThan)
+            ? _$$PersonImplPerFieldToJson.lastName(isLessThan)
             : null,
         isLessThanOrEqualTo: isLessThanOrEqualTo != null
-            ? _$$_PersonPerFieldToJson.lastName(isLessThanOrEqualTo)
+            ? _$$PersonImplPerFieldToJson.lastName(isLessThanOrEqualTo)
             : null,
         isGreaterThan: isGreaterThan != null
-            ? _$$_PersonPerFieldToJson.lastName(isGreaterThan)
+            ? _$$PersonImplPerFieldToJson.lastName(isGreaterThan)
             : null,
         isGreaterThanOrEqualTo: isGreaterThanOrEqualTo != null
-            ? _$$_PersonPerFieldToJson.lastName(isGreaterThanOrEqualTo)
+            ? _$$PersonImplPerFieldToJson.lastName(isGreaterThanOrEqualTo)
             : null,
         isNull: isNull,
-        whereIn: whereIn?.map((e) => _$$_PersonPerFieldToJson.lastName(e)),
+        whereIn: whereIn?.map((e) => _$$PersonImplPerFieldToJson.lastName(e)),
         whereNotIn:
-            whereNotIn?.map((e) => _$$_PersonPerFieldToJson.lastName(e)),
+            whereNotIn?.map((e) => _$$PersonImplPerFieldToJson.lastName(e)),
       ),
       $queryCursor: $queryCursor,
     );
@@ -734,7 +735,7 @@ class _$PersonQuery extends QueryReference<Person, PersonQuerySnapshot>
     PersonDocumentSnapshot? startAfterDocument,
   }) {
     final query = $referenceWithoutCursor
-        .orderBy(_$$_PersonFieldMap['firstName']!, descending: descending);
+        .orderBy(_$$PersonImplFieldMap['firstName']!, descending: descending);
     var queryCursor = $queryCursor;
 
     if (startAtDocument != null) {
@@ -806,7 +807,7 @@ class _$PersonQuery extends QueryReference<Person, PersonQuerySnapshot>
     PersonDocumentSnapshot? startAfterDocument,
   }) {
     final query = $referenceWithoutCursor
-        .orderBy(_$$_PersonFieldMap['lastName']!, descending: descending);
+        .orderBy(_$$PersonImplFieldMap['lastName']!, descending: descending);
     var queryCursor = $queryCursor;
 
     if (startAtDocument != null) {
@@ -1130,10 +1131,10 @@ class _$PublicRedirectedDocumentReference extends FirestoreDocumentReference<
     );
     final json = {
       if (value != _sentinel)
-        _$$PublicRedirected2FieldMap['value']!:
-            _$$PublicRedirected2PerFieldToJson.value(value as String),
+        _$$PublicRedirected2ImplFieldMap['value']!:
+            _$$PublicRedirected2ImplPerFieldToJson.value(value as String),
       if (valueFieldValue != null)
-        _$$PublicRedirected2FieldMap['value']!: valueFieldValue,
+        _$$PublicRedirected2ImplFieldMap['value']!: valueFieldValue,
     };
 
     return reference.update(json);
@@ -1150,10 +1151,10 @@ class _$PublicRedirectedDocumentReference extends FirestoreDocumentReference<
     );
     final json = {
       if (value != _sentinel)
-        _$$PublicRedirected2FieldMap['value']!:
-            _$$PublicRedirected2PerFieldToJson.value(value as String),
+        _$$PublicRedirected2ImplFieldMap['value']!:
+            _$$PublicRedirected2ImplPerFieldToJson.value(value as String),
       if (valueFieldValue != null)
-        _$$PublicRedirected2FieldMap['value']!: valueFieldValue,
+        _$$PublicRedirected2ImplFieldMap['value']!: valueFieldValue,
     };
 
     transaction.update(reference, json);
@@ -1487,30 +1488,31 @@ class _$PublicRedirectedQuery
     return _$PublicRedirectedQuery(
       _collection,
       $referenceWithoutCursor: $referenceWithoutCursor.where(
-        _$$PublicRedirected2FieldMap['value']!,
+        _$$PublicRedirected2ImplFieldMap['value']!,
         isEqualTo: isEqualTo != null
-            ? _$$PublicRedirected2PerFieldToJson.value(isEqualTo)
+            ? _$$PublicRedirected2ImplPerFieldToJson.value(isEqualTo)
             : null,
         isNotEqualTo: isNotEqualTo != null
-            ? _$$PublicRedirected2PerFieldToJson.value(isNotEqualTo)
+            ? _$$PublicRedirected2ImplPerFieldToJson.value(isNotEqualTo)
             : null,
         isLessThan: isLessThan != null
-            ? _$$PublicRedirected2PerFieldToJson.value(isLessThan)
+            ? _$$PublicRedirected2ImplPerFieldToJson.value(isLessThan)
             : null,
         isLessThanOrEqualTo: isLessThanOrEqualTo != null
-            ? _$$PublicRedirected2PerFieldToJson.value(isLessThanOrEqualTo)
+            ? _$$PublicRedirected2ImplPerFieldToJson.value(isLessThanOrEqualTo)
             : null,
         isGreaterThan: isGreaterThan != null
-            ? _$$PublicRedirected2PerFieldToJson.value(isGreaterThan)
+            ? _$$PublicRedirected2ImplPerFieldToJson.value(isGreaterThan)
             : null,
         isGreaterThanOrEqualTo: isGreaterThanOrEqualTo != null
-            ? _$$PublicRedirected2PerFieldToJson.value(isGreaterThanOrEqualTo)
+            ? _$$PublicRedirected2ImplPerFieldToJson
+                .value(isGreaterThanOrEqualTo)
             : null,
         isNull: isNull,
-        whereIn:
-            whereIn?.map((e) => _$$PublicRedirected2PerFieldToJson.value(e)),
-        whereNotIn:
-            whereNotIn?.map((e) => _$$PublicRedirected2PerFieldToJson.value(e)),
+        whereIn: whereIn
+            ?.map((e) => _$$PublicRedirected2ImplPerFieldToJson.value(e)),
+        whereNotIn: whereNotIn
+            ?.map((e) => _$$PublicRedirected2ImplPerFieldToJson.value(e)),
       ),
       $queryCursor: $queryCursor,
     );
@@ -1600,7 +1602,7 @@ class _$PublicRedirectedQuery
     PublicRedirectedDocumentSnapshot? startAfterDocument,
   }) {
     final query = $referenceWithoutCursor.orderBy(
-        _$$PublicRedirected2FieldMap['value']!,
+        _$$PublicRedirected2ImplFieldMap['value']!,
         descending: descending);
     var queryCursor = $queryCursor;
 
@@ -1764,45 +1766,48 @@ class PublicRedirectedQueryDocumentSnapshot
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_Person _$$_PersonFromJson(Map<String, dynamic> json) => _$_Person(
+_$PersonImpl _$$PersonImplFromJson(Map<String, dynamic> json) => _$PersonImpl(
       firstName: json['first_name'] as String,
       lastName: json['LAST_NAME'] as String,
     );
 
-const _$$_PersonFieldMap = <String, String>{
+const _$$PersonImplFieldMap = <String, String>{
   'firstName': 'first_name',
   'lastName': 'LAST_NAME',
 };
 
 // ignore: unused_element
-abstract class _$$_PersonPerFieldToJson {
+abstract class _$$PersonImplPerFieldToJson {
   // ignore: unused_element
   static Object? firstName(String instance) => instance;
   // ignore: unused_element
   static Object? lastName(String instance) => instance;
 }
 
-Map<String, dynamic> _$$_PersonToJson(_$_Person instance) => <String, dynamic>{
+Map<String, dynamic> _$$PersonImplToJson(_$PersonImpl instance) =>
+    <String, dynamic>{
       'first_name': instance.firstName,
       'LAST_NAME': instance.lastName,
     };
 
-_$PublicRedirected2 _$$PublicRedirected2FromJson(Map<String, dynamic> json) =>
-    _$PublicRedirected2(
+_$PublicRedirected2Impl _$$PublicRedirected2ImplFromJson(
+        Map<String, dynamic> json) =>
+    _$PublicRedirected2Impl(
       value: json['value'] as String,
     );
 
-const _$$PublicRedirected2FieldMap = <String, String>{
+const _$$PublicRedirected2ImplFieldMap = <String, String>{
   'value': 'value',
 };
 
 // ignore: unused_element
-abstract class _$$PublicRedirected2PerFieldToJson {
+abstract class _$$PublicRedirected2ImplPerFieldToJson {
   // ignore: unused_element
   static Object? value(String instance) => instance;
 }
 
-Map<String, dynamic> _$$PublicRedirected2ToJson(_$PublicRedirected2 instance) =>
+Map<String, dynamic> _$$PublicRedirected2ImplToJson(
+        _$PublicRedirected2Impl instance) =>
     <String, dynamic>{
       'value': instance.value,
     };