+++ packages/_flutterfire_internals/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.11
+
+ - Update a dependency to the latest release.
+
+## 1.3.10
+
+ - Update a dependency to the latest release.
+
 ## 1.3.9
 
  - Update a dependency to the latest release.