+++ packages/cloud_firestore/cloud_firestore_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.8.3
+
+ - Update a dependency to the latest release.
+
+## 3.8.2
+
+ - Update a dependency to the latest release.
+
 ## 3.8.1
 
  - Update a dependency to the latest release.