+++ packages/cloud_functions/cloud_functions_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.6.6
+
+ - Update a dependency to the latest release.
+
+## 4.6.5
+
+ - Update a dependency to the latest release.
+
 ## 4.6.4
 
  - Update a dependency to the latest release.