+++ packages/firebase_messaging/firebase_messaging/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 14.7.3
+
+ - Update a dependency to the latest release.
+
+## 14.7.2
+
+ - Update a dependency to the latest release.
+
 ## 14.7.1
 
  - Update a dependency to the latest release.