+++ packages/firebase_messaging/firebase_messaging_web/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.5.12
+
+ - Update a dependency to the latest release.
+
+## 3.5.11
+
+ - Update a dependency to the latest release.
+
 ## 3.5.10
 
  - Update a dependency to the latest release.