+++ packages/firebase_remote_config/firebase_remote_config/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.3.3
+
+ - Update a dependency to the latest release.
+
+## 4.3.2
+
+ - Update a dependency to the latest release.
+
 ## 4.3.1
 
  - Update a dependency to the latest release.